add routing in admin and conditional navbar rendering
