+++ client/src/components/MeritCertificate.jsx
@@ -80,7 +80,7 @@ const MeritCertificate = () => {
 
     return (
         <>
-            <div className='w-full bg-white rounded-md mt-3 flex flex-col justify-center'>
+            <div className='w-full bg-white rounded-md mt-3 flex flex-col justify-center shadow-[0_20px_20px_rgba(0,0,0,0.25)]'>
                 <h1 className='w-full text-center font-bold py-4 text-xl'>Add Merit Certificate</h1>
                 <div className="flex justify-center items-center gap-3 py-3">
                     <p className="font-bold text-md">Player Name :</p>
