+++ client/src/components/Navbar.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react'
+import React, { useEffect,useState } from 'react'
 import { NavLink, Routes, Route, Link } from 'react-router-dom'
 import '../App.css'
 import Home from "./Home.jsx"
@@ -15,7 +15,19 @@ import Contact from './Contact.jsx'
 
 const Navbar = () => {
 
- 
+    
+    const [isLoggedIn,setIsLoggedIn] = useState(false);
+
+    const getLoginState = () =>{
+        
+        if(document.cookie.split("loginState=")[1]){
+          setIsLoggedIn(true);
+        }
+    }
+
+    useEffect(()=>{
+        getLoginState();
+    },[])
 
 
 
@@ -30,9 +42,18 @@ const Navbar = () => {
               
                 <div className='w-auto links_container hidden md:flex justify-items-end items-center space-x-20 px-3  bg-transparent text-[.7rem] lg:text-lg xl:text-xl font-bold'>
                    
+                   {
+                    isLoggedIn
+                    ? 
+                    <></>
+                    :
+                    <>
                      <NavLink to="playerLogin" className="border border-white px-5 py-3 rounded-xl"> Login</NavLink>
                      <NavLink to="playerRegistor" className="border border-white px-5 py-3 rounded-xl"> Sign Up</NavLink>
                      
+                    </>
+                   }
+                    
                     </div>
 
             </nav>
