+++ client/src/components/Tournaments.jsx
@@ -48,7 +48,7 @@ const Tournaments = () => {
 
   return (
     <>
-      <form className="w-full bg-white rounded-md mt-3" onSubmit={addTournament}>
+      <form className="w-full bg-white rounded-md mt-3 shadow-[0_20px_20px_rgba(0,0,0,0.25)]" onSubmit={addTournament}>
         <h1 className="w-full text-center font-bold py-4 text-xl">
           Add New Tournament
         </h1>
